refactor(agent): agent runner architecture
